changes inspired by clang -Weverything

+++ src/global.h
@@ -19,8 +19,8 @@ enum {
 typedef struct {
     Instr* instr;
     long   phase;
-    double vol;
-    double rampvol;
+    float  vol;
+    float  rampvol;
     long   note_rate;
     int    key_state;
     int    age;
