+++ src/instr.h
@@ -1,4 +1,4 @@
 /* #include "global.h" */
 
-double instr_env_get(Note* note);
+float instr_env_get(Note* note);
 float instr_get(Note* note, long srate);
